backend/pyproject.toml
[project]
name = "synthedge-backend"
version = "0.1.0"
description = "SynthEdge Trading Terminal Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "asyncpg>=0.30.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "hyperliquid-python-sdk>=0.4.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
